requirements.txt
# Core PaddleOCR dependencies
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# PDF processing
pdf2image>=1.16.3
PyPDF2>=3.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI and visualization
click>=8.1.0
matplotlib>=3.7.0
colorama>=0.4.6
